+++ webtau-groovy-ast/src/main/groovy/com/twosigma/webtau/groovy/ast/ShouldAstTransformation.groovy
@@ -16,64 +16,22 @@
 
 package com.twosigma.webtau.groovy.ast
 
-import com.twosigma.webtau.Ddjt
+import groovy.transform.CompileStatic
 import org.codehaus.groovy.ast.ASTNode
 import org.codehaus.groovy.ast.ClassCodeExpressionTransformer
-import org.codehaus.groovy.ast.ClassNode
-import org.codehaus.groovy.ast.expr.ArgumentListExpression
-import org.codehaus.groovy.ast.expr.BinaryExpression
-import org.codehaus.groovy.ast.expr.ClosureExpression
-import org.codehaus.groovy.ast.expr.Expression
-import org.codehaus.groovy.ast.expr.MethodCallExpression
-import org.codehaus.groovy.ast.expr.PropertyExpression
-import org.codehaus.groovy.ast.expr.StaticMethodCallExpression
 import org.codehaus.groovy.control.CompilePhase
 import org.codehaus.groovy.control.SourceUnit
 import org.codehaus.groovy.transform.ASTTransformation
 import org.codehaus.groovy.transform.GroovyASTTransformation
 
+@CompileStatic
 @GroovyASTTransformation(phase=CompilePhase.SEMANTIC_ANALYSIS)
 class ShouldAstTransformation implements ASTTransformation {
+    public static final String SHOULD_BE_REPLACED_MESSAGE = "this method call should have been replaced with AST"
+
     @Override
     void visit(ASTNode[] astNodes, SourceUnit sourceUnit) {
-        ClassCodeExpressionTransformer transformer = new ClassCodeExpressionTransformer() {
-            @Override
-            protected SourceUnit getSourceUnit() {
-                return sourceUnit
-            }
-
-            @Override
-            Expression transform(Expression exp) {
-                if (exp instanceof ClosureExpression) {
-                    visitClosureExpression(exp)
-                }
-
-                if (! (exp instanceof BinaryExpression)) {
-                    return super.transform(exp)
-                }
-
-                BinaryExpression binaryExpression = exp as BinaryExpression
-                if (binaryExpression.operation.text != '==' ||
-                    ! (binaryExpression.leftExpression instanceof PropertyExpression)) {
-                    return super.transform(exp)
-                }
-
-                PropertyExpression leftExpression = binaryExpression.leftExpression as PropertyExpression
-                if (leftExpression.propertyAsString != 'should' &&
-                    leftExpression.propertyAsString != 'shouldNot' &&
-                    leftExpression.propertyAsString != 'waitTo' &&
-                    leftExpression.propertyAsString != 'waitToNot') {
-                    return super.transform(exp)
-                }
-
-                def methodCallExpression = new MethodCallExpression(leftExpression.objectExpression, leftExpression.property,
-                    new ArgumentListExpression(new StaticMethodCallExpression(
-                        new ClassNode(Ddjt), 'equal', binaryExpression.rightExpression)))
-
-                return methodCallExpression
-            }
-        }
-
+        ClassCodeExpressionTransformer transformer = new ShouldAstCodeTransformer()
         sourceUnit.AST.classes.each { transformer.visitClass(it) }
     }
 }
